requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
gtts>=2.5.0
SpeechRecognition>=3.10.0
audio-recorder-streamlit>=0.0.8
pandas>=2.0.0
